fix(rbac): deduplicate API route rules into single shared config
- Create lib/api-routes-config.js as the single source of truth
- Includes missing /api/upload/certificate and /api/achievements rules
- rbac-policy.js re-exports from shared config
- rbac.js imports from shared config instead of rbac-policy
- middleware.js imports from shared config, removing inline definitions
