+++ lib/rbac.js
@@ -1,66 +1,9 @@
 import { authenticateRequest } from "@/lib/error-handler";
 import { getUserProfile } from "@/lib/firebase-admin";
 import { ForbiddenError } from "@/lib/errors";
-import getApiRouteRule, { normalizeRoles } from "@/lib/rbac-policy";
+import getApiRouteRule, { normalizeRoles } from "@/lib/api-routes-config";
 
-export const PUBLIC_API_PATHS = new Set([
-  "/api/auth/csrf",
-  "/api/auth/reset-password",
-  "/api/health",
-]);
 
-const API_ROUTE_RULES = [
-  { pattern: /^\/api\/student(?:\/|$)/, roles: ["student", "admin"] },
-  { pattern: /^\/api\/teacher(?:\/|$)/, roles: ["teacher", "admin"] },
-  { pattern: /^\/api\/admin(?:\/|$)/, roles: ["admin"] },
-  { pattern: /^\/api\/institute(?:\/|$)/, roles: ["institute", "admin"] },
-  { pattern: /^\/api\/parent(?:\/|$)/, roles: ["parent", "admin"] },
-  {
-    pattern: /^\/api\/analytics\/attendance-risk(?:\/|$)/,
-    roles: ["teacher", "institute", "admin"],
-  },
-  {
-    pattern: /^\/api\/attendance\/settings(?:\/|$)/,
-    roles: ["teacher", "admin"],
-  },
-  { pattern: /^\/api\/attendance\/record(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/attendance\/sync(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/attendance\/validate-passcode(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/attendance\/heatmap(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/activities(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/auth\/cleanup(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/auth\/me(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/auth\/session(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/auth\/set-role(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/check-groq-config(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/complaints(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/conversations(?:\/|$)/, authOnly: true },
-  {
-    pattern: /^\/api\/flashcards(?:\/|$)/,
-    roles: ["student", "teacher", "admin"],
-  },
-  { pattern: /^\/api\/groq(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/images(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/labels(?:\/|$)/, roles: ["admin", "teacher", "student"] },
-  { pattern: /^\/api\/notifications(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/notifications\/seed(?:\/|$)/, roles: ["admin"] },
-  { pattern: /^\/api\/notices(?:\/|$)/, roles: ["teacher", "admin", "staff"] },
-  {
-    pattern: /^\/api\/productivity(?:\/|$)/,
-    roles: ["student", "teacher", "admin"],
-  },
-  { pattern: /^\/api\/settings(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/stats(?:\/|$)/, authOnly: true },
-  { pattern: /^\/api\/upload\/avatar(?:\/|$)/, authOnly: true },
-  {
-    pattern: /^\/api\/upload\/certificate(?:\/|$)/,
-    roles: ["teacher", "admin"],
-  },
-  {
-    pattern: /^\/api\/achievements(?:\/|$)/,
-    authOnly: true,
-  },
-];
 
 function isEmailVerified(payload) {
   return payload?.email_verified === true || payload?.emailVerified === true;
